[#1065] WeekEvents 순수 뷰와 표시 모델을 WidgetScenes 로 내린다
- `WeekEventsView` 와 `WeekEventsViewModel`·`WeekEventsRange` 가 `WidgetScenes` 로 갔다. `.sample(_:)` 이 쓰는 `EventOnWeek.dummy`·`DummyCalendarEvent` 도 함께 내려 갤러리가 표본을 그릴 수 있다
- `WeeksWithRange` 와 주 단위 계산 extension 은 확장에 남겼다. Provider 만 쓰는 조립용이다
- `WeekEventsRange` 에 `Sendable` 을 명시했다. 모듈을 넘으면서 public 타입의 자동 Sendable 추론이 사라져 TimelineProvider 의 `@Sendable` 클로저 캡처가 막혔다
