+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/WeekEventsWidget/WeekEventsWidgetTimelineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct WeekEventsWidgetTimelineProvider: TimelineProvider {
